requirements.txt
Package               Version
--------------------- -----------
allure-pytest         2.13.4
allure-python-commons 2.13.4
attrs                 23.2.0
axe-selenium-python   2.1.6
certifi               2024.2.2
charset-normalizer    3.3.2
h11                   0.14.0
idna                  3.6
iniconfig             2.0.0
outcome               1.3.0.post0
packaging             24.0
pillow                10.3.0
pip                   24.0
pixelmatch            0.3.0
pluggy                1.4.0
PySocks               1.7.1
pytest                8.1.1
pytest-check          2.3.1
pytest-dependency     0.6.0
pytest-dotenv         0.5.2
pytest-rerunfailures  14.0
python-dotenv         1.0.1
requests              2.31.0
selenium              4.19.0
setuptools            65.5.1
sniffio               1.3.1
sortedcontainers      2.4.0
trio                  0.25.0
trio-websocket        0.11.1
typing_extensions     4.10.0
urllib3               2.2.1
webdriver-manager     4.0.1
wheel                 0.38.4
wsproto               1.2.0

selenium~=4.19.0
pytest~=8.1.1
python-dotenv~=1.0.1
requests~=2.31.0
pillow~=10.3.0
pixelmatch~=0.3.0